+++ redux-counter/hook-counter/src/CounterContext.jsx
@@ -1,5 +1,6 @@
 /* eslint-disable react/prop-types */
-import { createContext, useReducer } from 'react'
+/* eslint-disable react-refresh/only-export-components */
+import { createContext, useReducer, useContext } from 'react'
 
 const counterReducer = (state, action) => {
   switch (action.type) {
@@ -15,6 +16,16 @@ const counterReducer = (state, action) => {
 }
 
 const CounterContext = createContext()
+
+export const useCounterValue = () => {
+  const counterAndDispatch = useContext(CounterContext)
+  return counterAndDispatch[0]
+}
+
+export const useCounterDispatch = () => {
+  const counterAndDispatch = useContext(CounterContext)
+  return counterAndDispatch[1]
+}
 
 export const CounterContextProvider = (props) => {
   const [counter, counterDispatch] = useReducer(counterReducer, 0)
