+++ redux-counter/hook-counter/src/components/Button.jsx
@@ -1,10 +1,8 @@
 /* eslint-disable react/prop-types */
-import { useContext } from 'react'
-import CounterContext from '../CounterContext'
+import { useCounterDispatch } from '../CounterContext'
 
 const Button = ({ type, label }) => {
-  // eslint-disable-next-line no-unused-vars
-  const [counter, dispatch] = useContext(CounterContext)
+  const dispatch = useCounterDispatch()
   return <button onClick={() => dispatch({ type })}>{label}</button>
 }
