+++ redux-counter/hook-counter/src/components/Display.jsx
@@ -1,8 +1,7 @@
-import { useContext } from 'react'
-import CounterContext from '../CounterContext'
+import { useCounterValue } from '../CounterContext'
 
 const Display = () => {
-  const [counter] = useContext(CounterContext)
+  const counter = useCounterValue()
   return <div>{counter}</div>
 }
